refactor(tokens): removing tokens/popper
