+++ packages/components/popper/__tests__/trigger.test.tsx
@@ -1,7 +1,7 @@
 import { nextTick, ref } from 'vue'
 import { mount } from '@vue/test-utils'
 import { afterEach, describe, expect, it, vi } from 'vitest'
-import { POPPER_INJECTION_KEY } from '@element-plus/tokens'
+import { POPPER_INJECTION_KEY } from '@element-plus/components/popper'
 import ElTrigger from '../src/trigger.vue'
 
 import type { VueWrapper } from '@vue/test-utils'
