+++ packages/components/popper/index.ts
@@ -14,5 +14,6 @@ export * from './src/popper'
 export * from './src/trigger'
 export * from './src/content'
 export * from './src/arrow'
+export * from './src/constants'
 
 export type { Placement, Options } from '@popperjs/core'
