+++ packages/components/popper/src/composables/use-content.ts
@@ -1,10 +1,10 @@
 import { computed, inject, onMounted, ref, unref, watch } from 'vue'
 import { isUndefined } from 'lodash-unified'
 import { usePopper } from '@element-plus/hooks'
-import { POPPER_INJECTION_KEY } from '@element-plus/tokens'
+import { POPPER_INJECTION_KEY } from '../constants'
 import { buildPopperOptions, unwrapMeasurableEl } from '../utils'
-import type { Modifier } from '@popperjs/core'
 
+import type { Modifier } from '@popperjs/core'
 import type { PartialOptions } from '@element-plus/hooks'
 import type { PopperContentProps } from '../content'
