+++ packages/components/popper/src/trigger.vue
@@ -17,8 +17,8 @@ import { isNil } from 'lodash-unified'
 import { unrefElement } from '@vueuse/core'
 import { ElOnlyChild } from '@element-plus/components/slot'
 import { useForwardRef } from '@element-plus/hooks'
-import { POPPER_INJECTION_KEY } from '@element-plus/tokens'
 import { isElement } from '@element-plus/utils'
+import { POPPER_INJECTION_KEY } from './constants'
 import { popperTriggerProps } from './trigger'
 
 import type { WatchStopHandle } from 'vue'
