+++ packages/tokens/index.ts
@@ -1,5 +1,4 @@
 export * from './experimental-features'
-export * from './popper'
 export * from './tooltip'
 export * from './tooltip-v2'
 export * from './date-picker'
